requirement.txt
streamlit>=1.27.0
numpy>=1.23.0
Pillow>=9.5.0
tensorflow>=2.11.0